+++ neoforge/src/main/java/cc/cassian/item_descriptions/client/neoforge/ItemDescriptionsNeoForge.java
@@ -39,7 +39,7 @@ public final class ItemDescriptionsNeoForge {
         //Only show tooltip if key is pressed or "always on" is enabled.
         if (showItemDescriptions()) {
             //Create and add tooltip. Tooltip will be wrapped.
-            List<Text> tooltip = createTooltip(findItemLoreKey(event.getItemStack()), true);
+            List<Text> tooltip = createTooltip(findItemLoreKey(event.getItemStack()), false);
             event.getToolTip().addAll(tooltip);
         }
     }